requirements.txt
streamlit==1.44.1
openai==1.17.0
neo4j==5.20.0
python-dotenv==1.0.1
pandas==2.2.1
tqdm==4.66.2
